refactor(storage): move test helper
